requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
black==24.1.1
flake8==7.0.0
drf-spectacular>=0.27.0
selenium>=4.18.1
webdriver-manager>=4.0.1 